+++ apps/example-vite-project/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @notjustcoders/example-vite-project
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @notjustcoders/one-logger-client-sdk@1.0.11
+
 ## 0.0.10
 
 ### Patch Changes