+++ packages/server-sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @notjustcoders/one-logger-server-sdk
 
+## 1.0.2
+
+### Patch Changes
+
+- refactor(config): update ports and cors configuration
+
 ## 1.0.1
 
 ### Patch Changes